requirements.txt
python-telegram-bot[job-queue]>=21.0
psutil
requests
Pillow>=10.0
fastapi>=0.100
uvicorn[standard]>=0.23
python-multipart>=0.0.6
